+++ src/liballoc/tests/lib.rs
@@ -9,7 +9,6 @@
 // except according to those terms.
 
 #![feature(allocator_api)]
-#![feature(alloc_system)]
 #![feature(box_syntax)]
 #![feature(drain_filter)]
 #![feature(exact_size_is_empty)]
@@ -20,7 +19,6 @@
 #![feature(unboxed_closures)]
 #![feature(repeat_generic_slice)]
 
-extern crate alloc_system;
 extern crate core;
 extern crate rand;
