+++ src/librustc_asan/Cargo.toml
@@ -15,6 +15,5 @@ cmake = "0.1.18"
 
 [dependencies]
 alloc = { path = "../liballoc" }
-alloc_system = { path = "../liballoc_system" }
 core = { path = "../libcore" }
 compiler_builtins = { path = "../rustc/compiler_builtins_shim" }
